// lib/quote-search-index.generated.ts
export type SearchableQuote = Quote & {
  episodeGuest: string;
  episodeUrl: string | null;
  topics: string[];
  transcriptText: string;
  normalized: {
    quote: string;
    speaker: string;
    episodeTitle: string;
    episodeGuest: string;
    episodeNumber: string;
    topics: string;
    transcript: string;
  };
};